Move fundingLoanHistory sync-schema to separate file
