+++ workers/loc.api/sync/schema/sync-schema/index.js
@@ -16,6 +16,7 @@ const statusMessages = require('./status-messages')
 const orders = require('./orders')
 const movements = require('./movements')
 const fundingOfferHistory = require('./funding-offer-history')
+const fundingLoanHistory = require('./funding-loan-history')
 
 const _methodCollMap = new Map([
   [SYNC_API_METHODS.LEDGERS, ledgers],
@@ -26,21 +27,7 @@ const _methodCollMap = new Map([
   [SYNC_API_METHODS.ORDERS, orders],
   [SYNC_API_METHODS.MOVEMENTS, movements],
   [SYNC_API_METHODS.FUNDING_OFFER_HISTORY, fundingOfferHistory],
-  [
-    SYNC_API_METHODS.FUNDING_LOAN_HISTORY,
-    {
-      name: ALLOWED_COLLS.FUNDING_LOAN_HISTORY,
-      maxLimit: 10000,
-      dateFieldName: 'mtsUpdate',
-      symbolFieldName: 'symbol',
-      sort: [['mtsUpdate', -1]],
-      hasNewData: false,
-      start: [],
-      isSyncRequiredAtLeastOnce: false,
-      type: COLLS_TYPES.INSERTABLE_ARRAY_OBJECTS,
-      model: getModelOf(TABLES_NAMES.FUNDING_LOAN_HISTORY)
-    }
-  ],
+  [SYNC_API_METHODS.FUNDING_LOAN_HISTORY, fundingLoanHistory],
   [
     SYNC_API_METHODS.FUNDING_CREDIT_HISTORY,
     {
